+++ R/stoxFunctionAttributes.R
@@ -103,11 +103,11 @@ stoxFunctionAttributes <- list(
             # The specification parameters for Baseline:
             RstoxBase::getFunctionArgumentHierarchyForSpcificationParameters(use = "Baseline", functionName = "ReportFunction"), 
             # The specification parameters for Boostrap:
-            RstoxBase::getFunctionArgumentHierarchyForSpcificationParameters(use = "Bootstrap", functionName = "BootstrapReportFunction"), 
-            functionParameterDefaults = list(
-                Percentages = c(5, 50, 95), 
-                GroupingVariables = c("Survey", "SpeciesCategory")
-            )
+            RstoxBase::getFunctionArgumentHierarchyForSpcificationParameters(use = "Bootstrap", functionName = "BootstrapReportFunction")
+        ), 
+        functionParameterDefaults = list(
+            Percentages = c(5, 50, 95), 
+            GroupingVariables = c("Survey", "SpeciesCategory")
         )
     ),